make fps display more accurate
GitPrivacy: IPRrj8sA1emx2dRUfDvhbPxUi9vlEGuY5JmlhSiv0QaLjICyrwCyZhK58tWkxaQ1D8j/08QElT8= jYu658YgMCIBYLzqcWxrtQm9cA0vOQ4LimzhCtFlHH580jFk4W8AOFkMhh1oL1zKLL9PvmqAsR0=

+++ src/main/scala/ScalaScape/components/CustomTypes.scala
@@ -2,5 +2,5 @@ package ScalaScape.components
 
 type WidthInColumns = Int
 type Between0And1   = Double
-type Milliseconds   = Long
+type Milliseconds   = Double
 type Seconds        = Double
